requirements.txt
asgiref==3.2.10
astroid==2.4.2
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
django_filter==22.1
tabula-py
wheel~=0.37.1
numpy~=1.23.0
setuptools~=60.2.0
distro~=1.7.0
Django~=4.0.6
colorama~=0.4.5
pandas~=1.4.3
python-dateutil~=2.8.2
django-filter~=22.1
tabula~=1.0.5